pyproject.toml
[tool.poetry]
name = "robolabs-api"
version = "0.1.0"
description = "Fetch and display information about invoices"
authors = ["Tomas <tomas.ribinskis@gmail.com>"]
readme = "README.md"
packages = [{include = "robolabs_api"}]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.3.2"
requests = "^2.31.0"
structlog = "^23.1.0"
setuptools = "^68.0.0"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120
